Alterada toda a base do projeto, incluindo carregameto estático de páginas, arquivos JS e CSS
