+++ core/Template.php
@@ -8,22 +8,23 @@ class Template
 	protected $img = 'imgs/';
 	protected $css = 'css/';
 	protected $js  = 'js/';
+	protected $pages = 'pages/';
 
 	public function construct()
 	{
 		$this->base = dirname(dirname(__FILE__));
 	}
 
-	public function getIncs($file = NULL)
+	public function getInc($file = NULL)
 	{
 		$file.='.php';
 
-		if (!file_exists($this->base . $file))
+		if (!file_exists($this->base . $this->inc . $file))
 		{
-			return json_encode(['error' => 'Arquivo não encontrado', 'file' => $file]);
+			return json_encode(['error' => 'Arquivo não encontrado ', 'file' => $file]);
 		}
 
-		include_once($this->base . $file);
+		include_once($this->base . $this->inc . $file);
 	}
 
 	public function getJS($file = NULL)
@@ -31,9 +32,9 @@ class Template
 		$file.='.js';
 
 		if (!file_exists($this->base . $this->assets . $this->js . $file))
-			return json_encode(['error' => 'Arquivo não encontrado', 'file' => $file]);
+			return json_encode(['error' => 'Arquivo não encontrado ', 'file' => $file]);
 		
-		return '<script src="'. $this->base . $this->js . $file . '"></script>';
+		return '<script src="'. $this->base . $this->assets . $this->js . $file . '"></script>';
 	}
 
 
@@ -42,9 +43,21 @@ class Template
 		$file.='.css';
 
 		if (!file_exists($this->base . $this->assets . $this->css . $file))
-			return json_encode(['error' => 'Arquivo não encontrado', 'file' => $file]);
+			return json_encode(['error' => 'Arquivo não encontrado ', 'file' => $file]);
 
-		return '<link rel="stylesheet" type="text/css" media="'.$media.'" href="'.$this->base . $this->assets . $this->css . $file.'" />';
+		return '<link rel="stylesheet" type="text/css" media="'.$media.'" href="'.$this->base . $this->assets . $this->css . $file.'" />' . "\n";
+	}
+
+	public function getPage($file = NULL)
+	{
+		$file.='.php';
+
+		if (!file_exists($this->base . $this->pages . $file))
+		{
+			return json_encode(['error' => 'Arquivo não encontrado ', 'file' => $file]);
+		}
+
+		include_once($this->base . $this->pages . $file);
 	}
 
 }
