+++ incs/footer.php
@@ -1,0 +1,2 @@
+</body>
+</html>
