+++ incs/menu.php
@@ -1,0 +1,32 @@
+    <!-- AREA DO MENU -->
+    <nav class="navbar navbar-default" style="margin-bottom:0;">
+	  <div class="container-fluid">
+	    <!-- Brand and toggle get grouped for better mobile display -->
+	    <div class="navbar-header">
+	      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
+	        <span class="sr-only">Toggle navigation</span>
+	        <span class="icon-bar"></span>
+	        <span class="icon-bar"></span>
+	        <span class="icon-bar"></span>
+	      </button>
+	      <a class="navbar-brand" href="#">
+	      	<img src="assets/img/logo-Combate-Mosquito.png" style="max-height:40px; margin-top:-10px; margin-right:30px;"/>
+	      </a>
+	    </div>
+
+	    <!-- Collect the nav links, forms, and other content for toggling -->
+	    <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+	      <ul class="nav navbar-nav">
+	        <li class="active"><a href="index.php">Mapa de Casos</a></li>
+	        <li><a href="?page=dicas">Dicas</a></li>
+	        <li><a href="?page=noticias">Notícias</a></li>
+	      </ul>
+	      
+	      <ul class="nav navbar-nav navbar-right">
+	        <li><a href="?page=contato">Contato</a></li>
+	       	<li><a href="?page=sobre">Sobre</a></li>
+	      </ul>
+	    </div><!-- /.navbar-collapse -->
+	  </div><!-- /.container-fluid -->
+	</nav>
+    <!-- AREA DO MENU -->
