--- noticias.php
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    
-    <link rel="stylesheet" type="text/css" media="all" href="css/bootstrap.min.css" />
-    <link rel="stylesheet" type="text/css" media="all" href="css/font-awesome.min.css" />
-    <link rel="stylesheet" type="text/css" media="all" href="css/theme.css" />
-    
-    <script src="js/jquery-2.1.1.min.js"></script>
-    <script src="js/bootstrap.min.js"></script>
-    
-    <script src="js/jquery.form-20140218.min.js"></script>    
-    <script src="js/script20160430.js"></script>
-
-    <title>DEV DAY 2016</title>
-    </head>
-
-    
-
-    <!-- AREA DO MENU -->
-    <nav class="navbar navbar-default" style="margin-bottom:0;">
-	  <div class="container-fluid">
-	    <!-- Brand and toggle get grouped for better mobile display -->
-	    <div class="navbar-header">
-	      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-	        <span class="sr-only">Toggle navigation</span>
-	        <span class="icon-bar"></span>
-	        <span class="icon-bar"></span>
-	        <span class="icon-bar"></span>
-	      </button>
-	      <a class="navbar-brand" href="#">
-	      	<img src="img/logo-Combate-Mosquito.png" style="max-height:40px; margin-top:-10px; margin-right:30px;"/>
-	      </a>
-	    </div>
-
-	    <!-- Collect the nav links, forms, and other content for toggling -->
-	    <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-	      <ul class="nav navbar-nav">
-	        <li><a href="index.html">Mapa de Casos</a></li>
-	        <li><a href="dicas.html">Dicas</a></li>
-	        <li class="active"><a  href="noticias.html">Notícias</a></li>
-	      </ul>
-	      
-	      <ul class="nav navbar-nav navbar-right">
-	        <li><a href="#">Contato</a></li>
-	       	<li><a href="sobre.html">Sobre</a></li>
-	      </ul>
-	    </div><!-- /.navbar-collapse -->
-	  </div><!-- /.container-fluid -->
-	</nav>
-    <!-- AREA DO MENU -->
-<div class="container">
-    <div class="space-sm"></div>
-    
-    <div class="text-justify">
-
-        <h3>Notícias</h3></br>
-    
-        <h4>Notícias 1</h4>
-        
-        <p>Cotéudo....</p></br>
-        
-        <h4>Notícias 2</h4>
-        
-        <p>Cotéudo....</p></br>
-        
-        <h4>Notícias 3</h4>
-        
-        <p>Cotéudo....</p>
-    </div>
-    
-    </div>
-</html>
