+++ pages/noticias.php
@@ -1,0 +1,22 @@
+
+<div class="container">
+    <div class="space-sm"></div>
+    
+    <div class="text-justify">
+
+        <h3>Notícias</h3></br>
+    
+        <h4>Notícias 1</h4>
+        
+        <p>Cotéudo....</p></br>
+        
+        <h4>Notícias 2</h4>
+        
+        <p>Cotéudo....</p></br>
+        
+        <h4>Notícias 3</h4>
+        
+        <p>Cotéudo....</p>
+    </div>
+    
+</div>
