+++ pages/sobre.php
@@ -1,0 +1,28 @@
+<div class="container">
+    <div class="space-sm"></div>
+    
+    <div class="text-justify">
+    
+        <div class="col-xs-11">
+        <h4>Sobre</h4>
+
+            <p>Projeto desenvolvido por estudantes das Faculdades Fatec de Mogi Mirim e Itapira, para o
+
+            campeonato de software Hackthon – DevDay 2016.
+
+            O projeto tem como desafio desenvolver um aplicativo que ajude a solucionar ou pelos menos
+
+            melhorar os problemas de comunicação entre população e prefeitura, no que se refere a
+
+            problemas de criadouro do mosquito da dengue, fornecendo meios tecnológicos para a
+
+            população, informar locais de foco e em contrapartida, ajudando os órgãos competentes a
+
+            localizar e combater os focos.
+
+            Também é disponível no site uma área com dicas e informativos, como campanhas e eventos
+
+            importantes para a população.</p>
+        </div>
+    </div>
+</div>
