--- sobre.php
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    
-    <link rel="stylesheet" type="text/css" media="all" href="css/bootstrap.min.css" />
-    <link rel="stylesheet" type="text/css" media="all" href="css/font-awesome.min.css" />
-    <link rel="stylesheet" type="text/css" media="all" href="css/theme.css" />
-    
-    <script src="js/jquery-2.1.1.min.js"></script>
-    <script src="js/bootstrap.min.js"></script>
-    
-    <script src="js/jquery.form-20140218.min.js"></script>    
-    <script src="js/script20160430.js"></script>
-
-    <title>DEV DAY 2016</title>
-    </head>
-
-    
-
-    <!-- AREA DO MENU -->
-    <nav class="navbar navbar-default" style="margin-bottom:0;">
-	  <div class="container-fluid">
-	    <!-- Brand and toggle get grouped for better mobile display -->
-	    <div class="navbar-header">
-	      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-	        <span class="sr-only">Toggle navigation</span>
-	        <span class="icon-bar"></span>
-	        <span class="icon-bar"></span>
-	        <span class="icon-bar"></span>
-	      </button>
-	      <a class="navbar-brand" href="#">
-	      	<img src="img/logo-Combate-Mosquito.png" style="max-height:40px; margin-top:-10px; margin-right:30px;"/>
-	      </a>
-	    </div>
-
-	    <!-- Collect the nav links, forms, and other content for toggling -->
-	    <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-	      <ul class="nav navbar-nav">
-	        <li><a href="index.html">Mapa de Casos</a></li>
-	        <li><a href="dicas.html">Dicas</a></li>
-	        <li><a href="noticias.html">Notícias</a></li>
-	      </ul>
-	      
-	      <ul class="nav navbar-nav navbar-right">
-	        <li><a href="#">Contato</a></li>
-	       	<li class="active"><a  href="sobre.html">Sobre</a></li>
-	      </ul>
-	    </div><!-- /.navbar-collapse -->
-	  </div><!-- /.container-fluid -->
-	</nav>
-    <!-- AREA DO MENU -->
-<div class="container">
-    <div class="space-sm"></div>
-    
-    <div class="text-justify">
-    
-        <div class="col-xs-11">
-        <h4>Sobre</h4>
-
-            <p>Projeto desenvolvido por estudantes das Faculdades Fatec de Mogi Mirim e Itapira, para o
-
-            campeonato de software Hackthon – DevDay 2016.
-
-            O projeto tem como desafio desenvolver um aplicativo que ajude a solucionar ou pelos menos
-
-            melhorar os problemas de comunicação entre população e prefeitura, no que se refere a
-
-            problemas de criadouro do mosquito da dengue, fornecendo meios tecnológicos para a
-
-            população, informar locais de foco e em contrapartida, ajudando os órgãos competentes a
-
-            localizar e combater os focos.
-
-            Também é disponível no site uma área com dicas e informativos, como campanhas e eventos
-
-            importantes para a população.</p>
-        </div>
-    </div>
-    </div>
-</html>
